+++ src/moulti/widgets/helpscreen.py
@@ -27,6 +27,8 @@ class HelpScreen(ModalScreen):
   [key]d[/]: switch between dark and light modes
   [key]c[/]: copy the contents of the current step
   [key]w[/]: same + preserve colors/include metadata (e.g. the question)
+  [key]/[/], [key]?[/]: forward/backward text search
+  [key]n[/], [key]N[/]: next/previous text search
   [key]q[/], [key]ctrl+c[/]: quit""".replace('[key]', '[bold #679cff on black]')
 
 	def compose(self) -> ComposeResult: